moved actionRouter and projectRouter into projects and actions folderss
