+++ projects/projectRouter.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const projectsDb = require("./data/helpers/projectModel");
+const projectsDb = require("../data/helpers/projectModel");
 
 const router = express.Router();
